requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
sqlmodel
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
asyncpg
python-multipart
psycopg2-binary
fastapi-mail
jinja2

# Testing
pytest
pytest-asyncio
httpx
aiosqlite
